maps beta diversity indicator based on spectral species distribution

maps beta diversity indicator based on spectral species distribution

map_beta_div(Input_Image_File, Output_Dir, window_size, TypePCA = "SPCA",
  nb_partitions = 20, nbclusters = 50, Nb_Units_Ordin = 2000,
  MinSun = 0.25, pcelim = 0.02, scaling = "PCO", FullRes = TRUE,
  LowRes = FALSE, nbCPU = 1, MaxRAM = 0.25)

Arguments

Input_Image_File
character. Path and name of the image to be processed.

Output_Dir
character. Output directory.

window_size
numeric. Dimensions of the spatial unit.

TypePCA
character. Type of PCA (PCA, SPCA, NLPCA...).

nb_partitions
numeric. Number of partitions (repetitions) to be computed then averaged.

nbclusters
numeric. Number of clusters defined in k-Means.

Nb_Units_Ordin
numeric. Maximum number of spatial units to be processed in NMDS. --> 1000 will be fast but may not capture important patterns if large area --> 4000 will be slow but may show better ability to capture landscape patterns

MinSun
numeric. Minimum proportion of sunlit pixels required to consider plot.

pcelim
numeric. Minimum contribution (in %) required for a spectral species

scaling
character. "PCO" or "NMDS".

FullRes
boolean.

LowRes
boolean.

nbCPU
numeric. Number of CPUs to use in parallel.

MaxRAM
numeric. MaxRAM maximum size of chunk in GB to limit RAM allocation when reading image file.

Value

None