maps functional diversity indicators based on prior selection of PCs

maps functional diversity indicators based on prior selection of PCs

map_functional_div(
  Original_Image_File,
  Functional_File = FALSE,
  Selected_Features = FALSE,
  Output_Dir,
  window_size,
  TypePCA = "SPCA",
  MinSun = 0.25,
  FullRes = FALSE,
  LowRes = TRUE,
  MapSTD = TRUE,
  nbCPU = 1,
  MaxRAM = 0.25,
  SmoothImage = TRUE
)

Arguments

Original_Image_File

character. Path and name of the original input image for biodivMapR.

Functional_File

character. Path and name of the image processed to be used to compute functional diversity --> can be PCA file with. if FALSE: using Original_Image_File

Selected_Features

numeric. Contains features to be used from Input_Image_File. using all if FALSE

Output_Dir

character. Output directory.

window_size

numeric. Size of spatial units (in pixels) to compute diversity.

TypePCA

character. Type of PCA (PCA, SPCA, NLPCA...).

MinSun

numeric. Minimum proportion of sunlit pixels required to consider plot.

FullRes

boolean. Full resolution.

LowRes

boolean. Low resolution.

MapSTD

boolean. map of standard deviation of the alpha diversity map (over repetitions)

nbCPU

numeric. Number of CPUs to use in parallel.

MaxRAM

numeric. MaxRAM maximum size of chunk in GB to limit RAM allocation when reading image file.

SmoothImage

boolean. set TRUE if you want smooting filter applied to resulting diversity rasters

Value

None