Performs PCA for all images and create PCA file with either all or a selection of PCs

Performs PCA for all images and create PCA file with either all or a selection of PCs

perform_PCA(ImPath, MaskPath, Output_Dir, Continuum_Removal = TRUE,
  TypePCA = "SPCA", NbPCs_To_Keep = FALSE, FilterPCA = FALSE,
  Excluded_WL = FALSE, nb_partitions = 20, nbCPU = 1,
  MaxRAM = 0.25)

Arguments

ImPath
character. Path of the image to be processed

MaskPath
character. Path of the mask corresponding to the image

Output_Dir
character. Path for output directory

Continuum_Removal
boolean. Set to TRUE if continuum removal should be applied

TypePCA
character. Type of PCA: choose either "PCA" or "SPCA"

NbPCs_To_Keep
numeric. number of components to ke saved in the PCA file. default = 30 if set to FALSE (or nb PC if <30)

FilterPCA
boolean. Set to TRUE if 2nd filtering based on PCA is required

Excluded_WL
numeric. Water Vapor Absorption domains (in nanometers, min and max WL). Can also be used to exclude spectific domains. dims = N x 2 (N = number of domains to be eliminated)

nb_partitions
numeric. Number of repetitions to estimate diversity from the raster (averaging repetitions).

nbCPU
numeric. Number fo CPUs in use.

MaxRAM
numeric. Maximum size of chunk in GB to limit RAM allocation when reading image file.

Value

list of paths corresponding to resulting PCA files